English search, Data cache, bug fixes
Added english search, touching the tableview from search now hides the
keyboard and does not select a row. User can now scroll the tableview
elements that are below the keyboard.

Now caches the previous data load if the networks is unavailable. Also
comes with a dict.json file that contains an initial dataset if the
user does not have network access on their first run.

Moved update button to the application support tab.

iOS 5 support, the project already worked on iOS5, just changed the
deployment target to reflect that.

Fixed the application support link being labeled as a search query.

Tab bar selection now updates when the view is reset.