Back Button, hides search bar and buttons on the support tab, other enhancements
Backbutton, allows the user to go back in their history without reseting

Search bar and tab bar buttons now hide on the application support tab.

added a function to clean the filter input

'search query' is now only shown on actual search queries and not on
taps to specific words

Added function to manage the graphical changes when switching tabs.